refactor : 패키지 정리 - 모델 분리(domain, service)
