+++ src/main/java/racing_game/model/domain/CarMovingStartegy.java
@@ -1,11 +1,11 @@
-package racing_game.model;
+package racing_game.model.domain;
 
 public interface CarMovingStartegy {
+
   /**
    * 차량이 움직일 수 있는지 판단
-   * 
-   * @param number : int, 판단에 필요한 정수
-   * @return
+   *
+   * @return boolean, 이동 가능 여부
    */
   boolean isCanMove();
 }
