+++ src/main/java/racing_game/utils/RacingGameUtils.java
@@ -1,17 +1,18 @@
-package racing_game.model;
+package racing_game.utils;
 
 import java.util.Random;
 
 public class RacingGameUtils {
-//  private static final int NUMBER_MOVEABLE_MIN = 4;
+
   private static final int NUMBER_RANDOM_BOUND = 10;
   private static final Random randomGenerator = new Random();
 
-  private RacingGameUtils() {}
+  private RacingGameUtils() {
+  }
 
   /**
    * 0에서 9 까지 무작위 정수 생성
-   * 
+   *
    * @return 생성된 숫자
    */
   public static int getRandomNumber() {
